added file delete files routes

+++ controller/file.controller.js
@@ -36,6 +36,10 @@ async function deleteFile(req, res) {
 
     const file = path.join(__dirname, "../assets/images", filename);
 
+    if (!data?.rows?.[0]?.doc.includes(`/assets/images/${filename}`)) {
+      return res.status(404).json({ message: "image not exist!" });
+    }
+
     if (fs.existsSync(file)) {
       fs.unlink(file, async (err) => {
         if (err) {
